@@ -13,9 +13,9 @@
         }
     },
     "require": {
-        "php": "^5.5 || ^7.0",
+        "php": "^5.6 || ^7.0",
         "zendframework/zend-escaper": "^2.5",
-        "zendframework/zend-stdlib": "^2.7 || ^3.0"
+        "zendframework/zend-stdlib": "^2.7 || ^3.1"
     },
     "require-dev": {
         "zendframework/zend-db": "^2.7",
@@ -23,7 +23,7 @@
         "zendframework/zend-http": "^2.5.4",
         "zendframework/zend-servicemanager": "^2.7.5 || ^3.0.3",
         "zendframework/zend-validator": "^2.6",
-        "phpunit/PHPUnit": "~4.0",
+        "phpunit/PHPUnit": "^6.0.8 || ^5.7.15",
         "psr/http-message": "^1.0",
         "zendframework/zend-coding-standard": "~1.0.0"
     },
@@ -35,12 +35,10 @@
         "zendframework/zend-servicemanager": "Zend\\ServiceManager component, for easily extending ExtensionManager implementations",
         "zendframework/zend-validator": "Zend\\Validator component, for validating email addresses used in Atom feeds and entries ehen using the Writer subcomponent"
     },
-    "minimum-stability": "dev",
-    "prefer-stable": true,
     "extra": {
         "branch-alias": {
-            "dev-master": "2.7-dev",
-            "dev-develop": "2.8-dev"
+            "dev-master": "2.8-dev",
+            "dev-develop": "2.9-dev"
         }
     },
     "autoload-dev": {
