@@ -9,11 +9,11 @@
 
 namespace ZendTest\Feed\Reader;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use ReflectionMethod;
 use Zend\Feed\Reader\FeedSet;
 
-class FeedSetTest extends PHPUnit_Framework_TestCase
+class FeedSetTest extends TestCase
 {
     /**
      * @var FeedSet
@@ -30,7 +30,7 @@ class FeedSetTest extends PHPUnit_Framework_TestCase
      */
     public function testAbsolutiseUri($link, $uri, $result)
     {
-        $method = new ReflectionMethod('Zend\Feed\Reader\FeedSet', 'absolutiseUri');
+        $method = new ReflectionMethod(FeedSet::class, 'absolutiseUri');
         $method->setAccessible(true);
 
         $this->assertEquals($result, $method->invoke($this->feedSet, $link, $uri));
